now the new vampire documents get pushed into an array that will be the argument for insertMany
