+++ models/vampires.js
@@ -1,58 +1,3 @@
-/////////////////////////////////////////////
-// Import Dependencies
-/////////////////////////////////////////////
-require("dotenv").config(); // Load ENV Variables
-const express = require("express"); // import express
-const morgan = require("morgan"); //import morgan
-const methodOverride = require("method-override");
-const mongoose = require("mongoose");
-const path = require("path")
-
-/////////////////////////////////////////////
-// Database Connection
-/////////////////////////////////////////////
-// Setup inputs for our connect function
-const DATABASE_URL = process.env.DATABASE_URL;
-const CONFIG = {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-};
-
-// Establish Connection
-mongoose.connect(DATABASE_URL, CONFIG);
-
-// Events for when connection opens/disconnects/errors
-const db = mongoose.connection
-    .on("open", () => console.log("Connected to Mongoose"))
-    .on("close", () => console.log("Disconnected from Mongoose"))
-    .on("error", (error) => console.log(error));
-
-////////////////////////////////////////////////
-// Models
-////////////////////////////////////////////////
-// pull schema and model from mongoose
-const { Schema, model } = mongoose;
-
-// make Vampires schema
-const vampiresSchema = new Schema(
-    {
-        name: { type: String, required: true },
-        title: String,
-        hair_color: { type: String, default: 'blonde' },
-        eye_color: String,
-        dob: Date,
-        loves: [String],
-        location: String,
-        gender: String,
-        victims: { type: Number, min: 0 }
-    },
-    { timestamps: true }
-);
-
-// make Vampire model
-const Vampire = model("Vampire", vampiresSchema);
-
-// Seed Data to insert into database
 const seedData = [
     {
         name: 'Count Chocula',
@@ -191,64 +136,5 @@ const seedData = [
         title: 'Osiris of Sewer Rats'
     }
 ]
-// Vampire.deleteMany({}).then((data) =>{
-Vampire.insertMany(seedData)
-    .then((data) => console.log('seedData inserted'))
-    .catch((error) => console.log(error))
-    // .finally(() => db.close())
-// })
 
-const Mark = {
-    name: 'Mark',
-    hair_color: 'red',
-    eye_color: 'blue',
-    dob: new Date(1986, 3, 27),
-    loves: ['climbing', 'coding', 'snacking'],
-    location: 'Boulder, Colorado, US',
-    gender: 'm',
-    title: 'Gerken puttin work in'
-}
-
-const John = {
-    name: 'John Johnson',
-    hair_color: 'brown',
-    eye_color: 'red',
-    dob: new Date(1914, 10, 3, 13, 12),
-    loves: ['being average', 'not making an impression', 'accounting'],
-    location: 'San Francisco, California, US',
-    gender: 'm',
-    title: 'The Most Forgettable Vampire Ever'
-}
-
-const Janet = {
-    name: 'Janet Bieber',
-    hair_color: 'blonde',
-    eye_color: 'brown',
-    dob: new Date(1998, 1, 20, 13, 12),
-    loves: ['draining the life out of fans', 'being famous'],
-    location: 'Quebec, Canada',
-    gender: 'f',
-    title: 'Pop Icon'
-}
-const Britney = {
-    name: 'Britney Spenser',
-    hair_color: 'blonde',
-    eye_color: 'green',
-    dob: new Date(1980, 6, 10),
-    loves: ['getting hit one more time'],
-    location: 'New York, New York, US',
-    gender: 'f',
-    title: 'Queen of Pop'
-}
-
-const createVampire = (name) => {
-    Vampire.create(name)
-        .then(data => console.log(data))
-        .catch(err => console.log(err))
-        // .finally(() => db.close())
-}
-
-createVampire(Mark)
-createVampire(John)
-createVampire(Janet)
-createVampire(Britney)
+module.exports = seedData
