+++ package.json
@@ -2,11 +2,11 @@
   "name": "mongoose-vampires",
   "version": "1.0.0",
   "description": "For this lab, you will be using some of the mongoose commands that you learned today and you will be **reading the documents** to find **new** queries to complete the following activities. Researching queries and implementing them is a big part of this lab!",
-  "main": "vampires.js",
+  "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node vampires.js",
-    "dev": "nodemon vampires.js"
+    "start": "node server.js",
+    "dev": "nodemon server.js"
   },
   "repository": {
     "type": "git",
